Bug 1580087. Fix 32bit * 32bit integer multiplies so they don't overflow in imagelib. r=aosmond
Differential Revision: https://phabricator.services.mozilla.com/D45303

UltraBlame original commit: 68594e925400aa41a448ff35ba346df6f3758704

+++ image/ClippedImage.cpp
@@ -512,7 +512,8 @@ nsIntSize ClippedImage::OptimalImageSizeForDest(const gfxSize& aDest,
 
     
     
-    gfxSize desiredSize(imgWidth * scale.width, imgHeight * scale.height);
+    gfxSize desiredSize(double(imgWidth) * scale.width,
+                        double(imgHeight) * scale.height);
     nsIntSize innerDesiredSize = InnerImage()->OptimalImageSizeForDest(
         desiredSize, aWhichFrame, aSamplingFilter, aFlags);
